# Updates in the `Horizon` extraction