--- a/seismiqb/__init__.py
+++ b/seismiqb/__init__.py
@@ -7,7 +7,7 @@ from .batch import SeismicCropBatch
 # Data entities
 from .field import Field, SyntheticField
 from .geometry import Geometry, array_to_segy, array_to_sgy
-from .labels import Horizon, HorizonPrototype, Fault, skeletonize
+from .labels import Horizon, HorizonExtractor, Fault, skeletonize
 from .metrics import HorizonMetrics, FaultsMetrics, FaciesMetrics
 from .samplers import GeometrySampler, HorizonSampler, FaultSampler, ConstantSampler, SeismicSampler
 from .grids import  BaseGrid, RegularGrid, ExtensionGrid, LocationsPotentialContainer