--- a/seismiqb/field/visualization.py
+++ b/seismiqb/field/visualization.py
@@ -347,7 +347,8 @@ class VisualizationMixin:
     CMAP_TO_ATTRIBUTE = {
         'Depths': ['full_matrix'],
         'Reds': ['spikes', 'quality_map', 'quality_grid'],
-        'Metric': ['metric']
+        'Metric': ['metric'],
+        'magma': ['mask_map']
     }
     ATTRIBUTE_TO_CMAP = {attr: cmap for cmap, attributes in CMAP_TO_ATTRIBUTE.items()
                          for attr in attributes}