--- a/seismiqb/functional/__init__.py
+++ b/seismiqb/functional/__init__.py
@@ -1,4 +1,4 @@
-""" !!. """
+""" General functions for working with seismic data. """
 # pylint: disable=wildcard-import
 from .base import *
 from .geologic_transforms import *