--- a/seismiqb/labels/__init__.py
+++ b/seismiqb/labels/__init__.py
@@ -1,3 +1,3 @@
 """ Labeled structures in a seismic volume. """
-from .horizon import Horizon
+from .horizon import Horizon, HorizonExtractor
 from .fault import Fault, skeletonize