--- a/seismiqb/labels/horizon/__init__.py
+++ b/seismiqb/labels/horizon/__init__.py
@@ -1,2 +1,3 @@
 """ Horizon class and mixins for POST-STACK data. """
 from .base import Horizon
+from .horizon_extractor import HorizonExtractor