--- a/seismiqb/labels/horizon/attributes.py
+++ b/seismiqb/labels/horizon/attributes.py
@@ -756,7 +756,7 @@ class AttributesMixin:
         grad_i = self.load_attribute('grad_i', on_full=True, dtype=np.float32, use_cache=False)
         grad_x = self.load_attribute('grad_x', on_full=True, dtype=np.float32, use_cache=False)
 
-        matrix = np.nanmax([grad_i, grad_x], axis=0)
+        matrix = np.abs(np.nanmax([grad_i, grad_x], axis=0))
         matrix[matrix == self.FILL_VALUE] = np.nan
         return matrix