**Updates in the `Horizon.from_mask` method:**

- [x] Add `mask_map` attribute evaluation into `Horizon.from_mask` (in some cases it can be a probabilities map)
- [x] Update method `Horizon.show()` for the 'probabilities' attribute
- [x] Refactor code using [connected_components](https://github.com/seung-lab/connected-components-3d/) library
- [x] bug fix in `Horizon.find_closest`

____
**Refactor horizon extraction for multiple horizons detection:**

- [x] Refactor code `from_mask`:
&nbsp; - disable extra `horizon.reset_cache` calls (multiple horizon extraction from peaked: **197 s -> 53 s** and **134 s -> 46 s**)
- [x] Refactor code `merge_into`:
&nbsp; - changed `horizon.bbox` view (10 prototypes extraction **414 s -> 294 s**)
&nbsp; - reduce list comprehensions (unused comprehension in **extraction.py** line 640 costs **43 s** from 287 s)
&nbsp; - reduce `np.ndarray` casting (notebook refactor with this in combination reduce timings: **227 s** -> **154 s**)
- [ ] Refactor code `verify_merge`:
&nbsp; - overlaps recalculation takes 15 s from 69 s; can reuse results from `merge_into` (not now)
&nbsp; - simplify method for lines merging
- [ ] Refactor code `overlap_merge`:
&nbsp; - simplify method for lines merging

____
____

Benchmarks for the `connected_components` library against previous implementation:

Comparison **only between** `skimage.label` vs `cc3d.connected_components`:

 - On the random generated data `np.random.binomial(n=1, p=0.3, size=size)`
![image](https://user-images.githubusercontent.com/21172482/185639871-e2755ca3-0577-4674-9f77-b2b9bb86b157.png)

 - On the real examples (`Horizon.from_mask` call):


**For one surface:**


`skimage` implementation:
![image](https://user-images.githubusercontent.com/21172482/185886541-b5fedae6-1e49-44d7-8d6a-1edc96acad93.png)


`cc3d` implementation (on the same data):
![image](https://user-images.githubusercontent.com/21172482/185886593-30913e54-2054-44be-84db-fddafd2e7092.png)



**For multiple surfaces:**

Time is in the format: **min:sec**

`skimage`: **6:35** + **7:53** &nbsp; _**VS**_ &nbsp; `cc3d`: **6:40** + **7:40**



**Result:** replacing `skimage.label` by `cc3d.connected_components`
____

A few words about comparison between `scipy.ndimage.find_objects` and `cc3d.each`:

The `find_objects` method returns slices, which are used for iterations filtering with too small labels.
This helps to save time by reducing `np.nonzero` calls, which are too slow.

The `cc3d.each` returns object masks and we need to evaluate filtering criterion, which is based on labels size.
Possible options are:
 - Call `find_objects` (Have no sense);
 - Take a `np.sum(object_mask)` (This is a weaker criterion, and method makes more `np.nonzero` calls);
 - Take `bounding_boxes` from `cc3d.statistics` (Statistics evaluation is too long - near to 2 minutes for each call, because the method evaluates other slow-computing statistics).

There are not an acceptable way to evaluate filtering criterion with the `cc3d.each` call. So, we didn't replace `scipy.ndimage.find_objects` by `cc3d.each`.